% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa.R
\docType{data}
\name{africa}
\alias{africa}
\title{Administrative Boundaries for 57 Countries within the African Continent}
\format{An object of class \code{SpatialPolygonsDataFrame} with 57 rows and 68 columns.}
\usage{
africa
}
\description{
\describe{
\item{ISO, ISO2, ISON}{}
\item{NAME_*}{Country name in a number of various dialects}
\item{VALIDFR, VALIDTO}{First / last year of nationhood}
\item{POP2000}{Population in Year 2000}
\item{SQKM}{Area in square kilometers}
\item{POPSQKM}{Population Density = (POP2000 / SQKM)}
\item{UNREGION1, UNREGION2}{}
\item{DEVELOPING}{}
\item{WBREGION}{World Bank Region}
\item{WBINCOME}{World Bank Income Class}
\item{WBDEBT}{World Bank Sovereign Debt Class}
}
}
\keyword{datasets}

