% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa.R
\docType{data}
\name{africa}
\alias{africa}
\title{Administrative Boundaries for Countries within the African Continent}
\format{An object of class \code{SpatialPolygonsDataFrame} with 54 rows and 65 columns.}
\usage{
africa
}
\description{
Source: \url{http://www.naturalearthdata.com/downloads/10m-cultural-vectors/}
}
\details{
\describe{
 \item{scalerank}{}
 \item{featurecla}{}
 \item{labelrank}{}
 \item{sovereignt}{}
 \item{sov_a3}{}
 \item{adm0_dif}{}
 \item{level}{}
 \item{type}{}
 \item{admin}{}
 \item{adm0_a3}{}
 \item{geou_dif}{}
 \item{geounit}{}
 \item{gu_a3}{}
 \item{su_dif}{}
 \item{subunit}{}
 \item{su_a3}{}
 \item{brk_diff}{}
 \item{name}{}
 \item{name_long}{}
 \item{brk_a3}{}
 \item{brk_name}{}
 \item{brk_group}{}
 \item{abbrev}{}
 \item{postal}{}
 \item{formal_en}{}
 \item{formal_fr}{}
 \item{note_adm0}{}
 \item{note_brk}{}
 \item{name_sort}{}
 \item{name_alt}{}
 \item{mapcolor7}{}
 \item{mapcolor8}{}
 \item{mapcolor9}{}
 \item{mapcolor13}{}
 \item{pop_est}{}
 \item{gdp_md_est}{}
 \item{pop_year}{}
 \item{lastcensus}{}
 \item{gdp_year}{}
 \item{economy}{}
 \item{income_grp}{}
 \item{wikipedia}{}
 \item{fips_10}{}
 \item{iso_a2}{}
 \item{iso_a3}{}
 \item{iso_n3}{}
 \item{un_a3}{}
 \item{wb_a2}{}
 \item{wb_a3}{}
 \item{woe_id}{}
 \item{adm0_a3_is}{}
 \item{adm0_a3_us}{}
 \item{adm0_a3_un}{}
 \item{adm0_a3_wb}{}
 \item{continent}{}
 \item{region_un}{}
 \item{subregion}{}
 \item{region_wb}{}
 \item{name_len}{}
 \item{long_len}{}
 \item{abbrev_len}{}
 \item{tiny}{}
 \item{homepart}{} 
}
}
\keyword{datasets}

